requirements.txt
asgiref==3.8.1
Django==5.2
djangorestframework==3.16.0
sqlparse==0.5.3
python-dotenv==1.1.0
django-autoslug==1.9.9
python-jose==3.4.0